requirements.txt
argparse
aiohttp
pyyaml
dnspython
pyfiglet
requests
